+++ src/main/kotlin/xyz/poeschl/bot/rosie/services/ConfigService.kt
@@ -51,8 +51,9 @@ class ConfigService(configFolder: Path) {
 
 @Serializable
 data class Config(
-    val telegramBotToken: String, val telegramBotUsername: String, val managingGroup: Long, val adminIds: List<Int> = emptyList(),
-    val cron: CronConfig = CronConfig(), val locationPoll: LocationPollConfig = LocationPollConfig(), val knownUsers: MutableMap<Int, String> = mutableMapOf()
+    val telegramBotToken: String, val telegramBotUsername: String, val managingGroup: Long, val adminIds: List<Long> = emptyList(),
+    val cron: CronConfig = CronConfig(), val locationPoll: LocationPollConfig = LocationPollConfig(), val knownUsers: MutableMap<Long,
+            String> = mutableMapOf()
 )
 
 @Serializable
